+++ src/swganh/messages/update_transform_with_parent_message.h
@@ -46,7 +46,7 @@ struct UpdateTransformWithParentMessage : public swganh::messages::BaseSwgMessag
         buffer.write<int16_t>(position.y * 8.0f + 0.5f);
         buffer.write<int16_t>(position.z * 8.0f + 0.5f);
         buffer.write(update_counter);
-        buffer.write(0);
+        buffer.write(static_cast<uint8_t>(0));
         buffer.write(heading);
     }
